informants/magnetometer: Add I2C Magnetometer code for comms testing
This is an initial commit of my rough first implementation of the driver
for the IST8310 magnetometer. It's essentially just another go at what
I'd already tried and failed at. The entire purpose is to serve as a
proof-of-concept and that we *can* actually use I2C so that when we work
on integrating other projects, we can be sure that the dev-board side is
working.